Reorganized the project structure and comments
