+++ clitoolsboilerplate/tools/demo.py
@@ -1,20 +1,27 @@
-# PYTHON_BOILER_PLATE Base Tool
-# BaseTool definition for command line tools.
+"""
+PYTHON_BOILER_PLATE Demo Tool
 
+An example of using the `BaseTool` class to create a command line tool.
+"""
 # MARK: Imports
 import argparse
 from typing import Optional
 
+from .baseTool import BaseTool
 from ..config import ConfigManager
 
 # MARK: Classes
-class BaseTool:
+class DemoTool(BaseTool):
     """
-    Base class for command line tools.
+    Demo tool for command line tools.
     """
     # Constants
-    TOOL_NAME = "base_tool"
-    TOOL_HELP = "Base tool for command line tools."
+    TOOL_NAME = "demo"
+    TOOL_HELP = "Demo tool for command line tools."
+
+    # Initializer
+    def __init__(self, isA: bool):
+        self.isA = isA
 
     # CLI Functions
     @staticmethod
@@ -25,25 +32,34 @@ class BaseTool:
         parser: The parser to apply the arguments to.
         config: The config manager to use for the tool or `None` if not present.
         """
-        raise NotImplementedError("setupParser() must be implemented in the subclass.")
+        # Add a demo argument
+        parser.add_argument(
+            "-d",
+            "--demo",
+            action="store_true",
+            help="Run the demo."
+        )
 
     @classmethod
-    def fromArgs(cls, args: argparse.Namespace, config: Optional[ConfigManager]) -> "BaseTool":
+    def fromArgs(cls, args: argparse.Namespace, config: Optional[ConfigManager]) -> "DemoTool":
         """
         Creates an instance of this tool from the given `args`.
 
         args: The parser arguments to create the tool from.
-        config: The config manager to use for the tool or `None` if not present.
+        config: The config manager to use for the tool.
 
         Returns an instance of this tool.
         """
-        raise NotImplementedError("fromArgs() must be implemented in the subclass.")
+        return cls(args.demo)
 
     def _run(self, args: argparse.Namespace, config: Optional[ConfigManager]):
         """
         Runs the tool as configured by the CLI.
 
         args: The parser arguments to create the tool from.
-        config: The config manager to use for the tool or `None` if not present.
+        config: The config manager to use for the tool.
         """
-        raise NotImplementedError("_run() must be implemented in the subclass.")
+        print("Running demo tool!")
+        print(f"Is A: {self.isA}")
+        print(f"Args: {args}")
+        print(f"Config: {config}")
